17-02-22 add ft_putnbr_unsigned_fd to libft

+++ test.c
@@ -6,7 +6,7 @@
 /*   By: hbourgeo <hbourgeo@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/15 08:23:29 by hbourgeo          #+#    #+#             */
-/*   Updated: 2022/02/18 08:28:08 by hbourgeo         ###   ########.fr       */
+/*   Updated: 2022/02/18 08:44:20 by hbourgeo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -197,10 +197,10 @@ int ft_printf_p (va_list arg)
 	void	*add;
 
 	size = 0;
-	add = (unsigned long long)va_arg(arg, void *);
+	add = va_arg(arg, void *);
 	size += ft_putstr_fd("0x", 1);
 	// size = ft_convert_hex(add);
-	size += ft_putnbr_base_fd (add, "0123456789abcdef", 1);
+	size += ft_putnbr_base_fd ((unsigned long long int)add, "0123456789abcdef", 1);
 	return(size);
 }
